+++ eyes.selenium.java/src/main/java/com/applitools/eyes/selenium/capture/DomCapture.java
@@ -4,10 +4,12 @@ import com.applitools.connectivity.ServerConnector;
 import com.applitools.eyes.*;
 import com.applitools.eyes.positioning.PositionMemento;
 import com.applitools.eyes.positioning.PositionProvider;
+import com.applitools.eyes.selenium.EyesSeleniumUtils;
 import com.applitools.eyes.selenium.SeleniumEyes;
 import com.applitools.eyes.selenium.frames.FrameChain;
-import com.applitools.eyes.selenium.wrappers.EyesTargetLocator;
+import com.applitools.eyes.selenium.rendering.VisualGridEyes;
 import com.applitools.eyes.selenium.wrappers.EyesSeleniumDriver;
+import com.applitools.eyes.selenium.wrappers.EyesTargetLocator;
 import com.applitools.eyes.visualgrid.model.RGridResource;
 import com.applitools.utils.EfficientStringReplace;
 import com.applitools.utils.GeneralUtils;
@@ -15,41 +17,20 @@ import org.openqa.selenium.By;
 import org.openqa.selenium.WebElement;
 
 import java.io.IOException;
-import java.net.*;
+import java.net.URI;
 import java.util.*;
 import java.util.concurrent.Phaser;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.TimeoutException;
-import java.util.concurrent.atomic.AtomicBoolean;
 
 public class DomCapture {
-
-    private class TimeoutTask extends TimerTask {
-        @Override
-        public void run() {
-            logger.verbose("Check Timer timeout");
-            isCheckTimerTimedOut.set(true);
-        }
-    }
-
-    private static String CAPTURE_FRAME_SCRIPT;
-
-    private static String CAPTURE_FRAME_SCRIPT_FOR_IE;
+    private final String CAPTURE_DOM;
+    private final String CAPTURE_DOM_FOR_IE;
+    private final String POLL_RESULT;
+    private final String POLL_RESULT_FOR_IE;
 
     private final Phaser cssPhaser = new Phaser(); // Phaser for syncing all callbacks on a single Frame
 
-    static {
-        try {
-            CAPTURE_FRAME_SCRIPT = GeneralUtils.readToEnd(DomCapture.class.getResourceAsStream("/captureDomAndPoll.js"));
-            CAPTURE_FRAME_SCRIPT += "return __captureDomAndPoll();";
-            CAPTURE_FRAME_SCRIPT_FOR_IE = GeneralUtils.readToEnd(DomCapture.class.getResourceAsStream("/captureDomAndPollForIE.js"));
-            CAPTURE_FRAME_SCRIPT_FOR_IE += "return __captureDomAndPollForIE();";
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-    }
-
-    private static final long DOM_EXTRACTION_TIMEOUT = 5 * 60 * 1000;
     private static ServerConnector serverConnector = null;
     private final EyesSeleniumDriver driver;
     private final Logger logger;
@@ -57,7 +38,6 @@ public class DomCapture {
     private String cssEndToken;
     private final Map<String, CssTreeNode> cssNodesToReplace = Collections.synchronizedMap(new HashMap<String, CssTreeNode>());
     private boolean shouldWaitForPhaser = false;
-    private final AtomicBoolean isCheckTimerTimedOut = new AtomicBoolean(false);
 
     private final UserAgent userAgent;
 
@@ -66,6 +46,15 @@ public class DomCapture {
         logger = eyes.getLogger();
         driver = (EyesSeleniumDriver) eyes.getDriver();
         userAgent = eyes.getUserAgent();
+
+        try {
+            CAPTURE_DOM = GeneralUtils.readToEnd(DomCapture.class.getResourceAsStream("/dom-capture/dist/captureDomAndPoll.js"));
+            CAPTURE_DOM_FOR_IE = GeneralUtils.readToEnd(DomCapture.class.getResourceAsStream("/dom-capture/dist/captureDomAndPollForIE.js"));
+            POLL_RESULT = GeneralUtils.readToEnd(VisualGridEyes.class.getResourceAsStream("/dom-capture/dist/pollResult.js"));
+            POLL_RESULT_FOR_IE = GeneralUtils.readToEnd(VisualGridEyes.class.getResourceAsStream("/dom-capture/dist/pollResultForIE.js"));
+        } catch (IOException e) {
+            throw new EyesException("Failed getting resources for dom scripts", e);
+        }
     }
 
     public String getPageDom(PositionProvider positionProvider) {
@@ -80,7 +69,7 @@ public class DomCapture {
 
         try {
             if (shouldWaitForPhaser) {
-                cssPhaser.awaitAdvanceInterruptibly(0, 30, TimeUnit.SECONDS);
+                cssPhaser.awaitAdvanceInterruptibly(0, 60, TimeUnit.SECONDS);
             }
         } catch (InterruptedException | TimeoutException e) {
             GeneralUtils.logExceptionStackTrace(logger, e);
@@ -98,68 +87,34 @@ public class DomCapture {
 
     private String getFrameDom(String baseUrl) {
         logger.verbose("Trying to get DOM from driver");
-        Timer timer = new Timer(true);
-        timer.schedule(new TimeoutTask(), DOM_EXTRACTION_TIMEOUT);
+        String domScript = userAgent.isInternetExplorer() ? CAPTURE_DOM_FOR_IE : CAPTURE_DOM;
+        String pollingScript = userAgent.isInternetExplorer() ? POLL_RESULT_FOR_IE : POLL_RESULT;
+
+        List<String> missingCssList = new ArrayList<>();
+        List<String> missingFramesList = new ArrayList<>();
+        List<String> data = new ArrayList<>();
+        Separators separators;
         try {
-            isCheckTimerTimedOut.set(false);
-            String resultAsString;
-            ScriptResponse.Status status = null;
-            ScriptResponse scriptResponse = null;
-            do {
-                if (userAgent.isInternetExplorer()) {
-                    resultAsString = (String) this.driver.executeScript(CAPTURE_FRAME_SCRIPT_FOR_IE);
-                } else {
-                    resultAsString = (String) this.driver.executeScript(CAPTURE_FRAME_SCRIPT);
-                }
+            String scriptResult = EyesSeleniumUtils.runDomScript(logger, driver, userAgent, domScript, null, pollingScript);
+            scriptResult = GeneralUtils.parseJsonToObject(scriptResult, String.class);
+            separators = parseScriptResult(scriptResult, missingCssList, missingFramesList, data);
+        } catch (Exception e) {
+            throw new EyesException("Failed running dom capture script", e);
+        }
 
-                try {
-                    scriptResponse = GeneralUtils.parseJsonToObject(resultAsString, ScriptResponse.class);
-                    status = scriptResponse.getStatus();
-                } catch (IOException e) {
-                    GeneralUtils.logExceptionStackTrace(logger, e);
-                }
-                Thread.sleep(200);
+        cssStartToken = separators.cssStartToken;
+        cssEndToken = separators.cssEndToken;
 
-            } while (status == ScriptResponse.Status.WIP && !isCheckTimerTimedOut.get());
-            timer.cancel();
+        fetchCssFiles(baseUrl, missingCssList, null);
 
-            if (scriptResponse == null || status == ScriptResponse.Status.ERROR) {
-                String errorMessage = scriptResponse == null ? "Script response is null" : scriptResponse.getError();
-                throw new EyesException("DomCapture Error: " + errorMessage);
-            }
-
-            if (isCheckTimerTimedOut.get()) {
-                throw new EyesException("DomCapture Timed out");
-            }
-            String executeScripString = scriptResponse.getValue();
-
-            List<String> missingCssList = new ArrayList<>();
-            List<String> missingFramesList = new ArrayList<>();
-            List<String> data = new ArrayList<>();
-
-            Separators separators = parseScriptResult(executeScripString, missingCssList, missingFramesList, data);
-            cssStartToken = separators.cssStartToken;
-            cssEndToken = separators.cssEndToken;
-
-            fetchCssFiles(baseUrl, missingCssList, null);
-
-            Map<String, String> framesData = new HashMap<>();
-            try {
-                framesData = recurseFrames(missingFramesList);
-            } catch (Exception e) {
-                GeneralUtils.logExceptionStackTrace(logger, e);
-            }
-
-
-            return EfficientStringReplace.efficientStringReplace(
-                    separators.iframeStartToken, separators.iframeEndToken, data.get(0), framesData);
-        } catch (InterruptedException e) {
+        Map<String, String> framesData = new HashMap<>();
+        try {
+            framesData = recurseFrames(missingFramesList);
+        } catch (Exception e) {
             GeneralUtils.logExceptionStackTrace(logger, e);
         }
-        finally {
-            timer.cancel();
-        }
-        return "";
+
+        return EfficientStringReplace.efficientStringReplace(separators.iframeStartToken, separators.iframeEndToken, data.get(0), framesData);
     }
 
     private Separators parseScriptResult(String scriptResult, List<String> missingCssList, List<String> missingFramesList, List<String> data) {
@@ -185,7 +140,7 @@ public class DomCapture {
             logger.verbose("missing css count: " + missingCssList.size());
             logger.verbose("missing frames count: " + missingFramesList.size());
         } catch (IOException e) {
-            e.printStackTrace();
+            GeneralUtils.logExceptionStackTrace(logger, e);
         }
         shouldWaitForPhaser |= !missingCssList.isEmpty();
         return separators;